requirements.txt
astroid==2.11.7
cachelib==0.9.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dill==0.3.5.1
Flask==2.1.3
Flask-Session==0.4.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
pyee==9.0.4
pylint==2.14.4
python-dotenv==0.20.0
requests==2.28.1
slack-bolt==1.14.1
slack-sdk==3.17.2
slackeventsapi==3.0.1
tomli==2.0.1
tomlkit==0.11.1
typing-extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.1
